feat: add 'copy to clipboard' button to schema errors
